Split MCP stdio codec and result mapping
