+++ docs/DECISIONS.md
@@ -199,7 +199,7 @@
 - Review action run planning lives in `WorkspaceReviewActionToolCallPlanner.swift`, not as extensions on review surface values or inline in `WorkspaceModel.swift`. The planner owns the canonical file/hunk stage and restore `host.git.*` argument shape, the mandatory follow-up `host.git.diff` refresh call, and final status derivation from action plus refresh results. The workspace model owns only executing the planned calls, recording tool cards, persistence, and top-bar application.
 - Production Swift sources should avoid `try!`, `as!`, and force unwraps. Startup, OAuth sign-in, local project scanning, and extension discovery must fail through typed errors or ignored unsafe records rather than crashing the process. `ProjectExtensionManifestLoader` normalizes extension directories by safe path components and skips unsafe custom directories without blocking later valid directories.
 - Agent tool execution override composition lives in `WorkspaceToolExecutionOverrideCombiner.swift`, not in `WorkspaceModel.swift`. The combiner owns the precedence order Plan, Remote Project, Browser, Computer Use, Memory, MCP, while the workspace model only creates each optional executor and assigns the combined override to the runner.
-- MCP extension surface state lives in `QuillCodeMCPSurface.swift`, MCP tool/resource/prompt JSON request parsing lives in `WorkspaceMCPRequests.swift`, MCP subprocess lifecycle and session routing live in `WorkspaceMCPRuntime.swift`, concrete stdio launch/prober construction lives behind `WorkspaceMCPServerLaunching` in `WorkspaceMCPServerLauncher.swift`, and Ready-server dynamic tool catalog generation lives in `WorkspaceMCPToolCatalog.swift`. `WorkspaceModel` owns selected-project manifest lookup, UI notices, top-bar status, and state application; surface labels, probe-summary compatibility, request aliases, nested-argument normalization, process/session launch seams, process handles, and catalog text stay in focused helpers with parity gates.
+- MCP extension surface state lives in `QuillCodeMCPSurface.swift`, MCP tool/resource/prompt JSON request parsing lives in `WorkspaceMCPRequests.swift`, MCP subprocess lifecycle and session routing live in `WorkspaceMCPRuntime.swift`, concrete stdio launch/prober construction lives behind `WorkspaceMCPServerLaunching` in `WorkspaceMCPServerLauncher.swift`, stdio frame parsing lives in `MCPStdioMessageCodec.swift`, MCP stdio response/result mapping lives in `MCPStdioResultMapper.swift`, and Ready-server dynamic tool catalog generation lives in `WorkspaceMCPToolCatalog.swift`. `WorkspaceModel` owns selected-project manifest lookup, UI notices, top-bar status, and state application; surface labels, probe-summary compatibility, request aliases, nested-argument normalization, process/session launch seams, process handles, stdio framing, result mapping, and catalog text stay in focused helpers with parity gates.
 - Workspace command IDs reduce through `WorkspaceCommandPlan` and execute through `WorkspaceCommandPlanExecutor` before touching `WorkspaceModel` side effects. The planner owns prefix parsing, draft-prefill mapping, static action mapping, automation quick recurrence parsing, and canonical command-ID-to-tool-name mapping. The executor owns the parsed-plan switch, while `WorkspaceModel` owns the underlying mutations and tool dispatch helpers. This keeps command palette rows, slash-command templates, automation row actions, MCP lifecycle actions, memory actions, and git shortcuts from drifting as the command set grows.
 - Workspace command actions reduce through `WorkspaceCommandActionPlanner` before touching `WorkspaceModel` side effects. The planner owns selected project/thread preconditions, rename draft construction, sidebar bulk command mapping, and action-to-effect routing. `WorkspaceModel` should execute typed `WorkspaceCommandActionEffect` values and keep actor-bound persistence, top-bar refresh, and store synchronization rather than growing another selected-state switch.
 - Command palette surface behavior lives in `WorkspaceCommandPaletteSurface.swift`. Command records, top-bar overflow projection, automation and Computer Use command factories, category ordering, slash/action query scoping, and ranking/scoring stay beside the command contract instead of inside the aggregate `WorkspaceSurface.swift` payload.
